% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{summary_plots}
\alias{summary_plots}
\title{Generate Summary Figures}
\usage{
summary_plots(
  effect_size_dataframe,
  Small_Sample_Correction = TRUE,
  Significance_Level = 0.05
)
}
\arguments{
\item{effect_size_dataframe}{Output from the \emph{classify_interactions} function.}

\item{Small_Sample_Correction}{Whether the correction for small sample sizes should be enacted
(TRUE or FALSE; default is TRUE) \emph{Note that if the multiplicative null model was implemented,
this parameter is not used and can be ignored. If the additive null model was implemented, then this
parameter should be assigned the same value as in that analysis}.}

\item{Significance_Level}{The value of alpha for which confidence intervals are calculated
(numeric, between 0 and 1; default is 0.05) \emph{Note that if the multiplicative null model was implemented,
this parameter is not used and can be ignored. If the additive null model was implemented, then this
parameter should be assigned the same value as in that analysis}.}
}
\value{
The function returns a series of figures each of which is outlined above.
}
\description{
Using the output from \emph{classify_interactions} function, summary figures can be created using this function.
}
\details{
The figures include:

 a) The proportions of the different interaction classifications from the dataset

 b) Median sample sizes plotted against effect size (different interaction classifications are highlighted).
    Where the additive null model was used in the analysis, lines for critical effect sizes are plotted (see \emph{critical_effect_size_additive} function).

 c) Density of different median sample sizes.

 d) Inverse of effect size variance plotted against effect size (i.e., one iteration of a funnel plot).

 e) Effect size standard error (i.e., the square root of the effect size variance) plotted against effect size (i.e., one iteration of a funnel plot)).

 Note that c - e) are most useful for researchers conducting a meta-analysis.
}
\examples{
classify_interactions(effect_size_dataframe = df,
                     Small_Sample_Correction = TRUE,
                     remove_directionality = TRUE)

classify_interactions(effect_size_dataframe = df)

}
