% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_interactions.R
\name{classify_interactions}
\alias{classify_interactions}
\title{Classify Interactions}
\usage{
classify_interactions(
  effect_size_dataframe,
  assign_reversals = TRUE,
  remove_directionality = FALSE
)
}
\arguments{
\item{effect_size_dataframe}{Output from either the \code{\link{effect_size_additive}} or \code{\link{effect_size_multiplicative}} functions.}

\item{assign_reversals}{Specify whether reversals should be distinguished from antagonisms (see \emph{Jackson et al. (2016)}
or \emph{Orr et al. (2020)}) (TRUE or FALSE; default = TRUE)}

\item{remove_directionality}{Specify whether directionality should be removed from the effect sizes.
This is most pertinent where subsequent analyses will involve conducting a formal meta-analysis (set value to TRUE).
For examples see \emph{Burgess et al. (2021)}; \emph{Jackson et al. (2016)}; \emph{Seifert et al. (2020)}; \emph{Siviter et al. (2021)}
(TRUE or FALSE; default = FALSE)}
}
\value{
The function returns a dataset containing
i)   the same results as from either the \emph{effect_size_additive} or \emph{effect_size_multiplicative} functions.

ii)  the classification of any interactions (see below)

iii) if remove_directionality = TRUE; effect sizes, and confidence intervals will be adjusted (see below)

---

---

For ii), interactions are classified in the following method.

 a) if the confidence intervals for any effect size overlap zero then the interaction is assigned a \emph{null} classification.

 This is analagous to the classification of \emph{additive} which other studies (e.g., \emph{Jackson et al. (2016)}) may use.

 Note that this is given precedence over the process described in b).


 ---

 b) if the confidence intervals for any effect size do not overlap zero then \emph{Expected} and \emph{Observed} effects are calculated.

 Where the additive null model has been used:
 \deqn{Expected = X_a + X_b - 2 * X_c}
 \deqn{Observed = X_i - X_c}

 Where the multiplicative null model has been used:
 \deqn{Expected = ln(X_a) + ln(X_b) - 2 * ln(X_c)}
 \deqn{Observed = ln(X_i) - ln(X_c)}

 Here \emph{X_c}, \emph{X_a}, \emph{X_b}, and \emph{X_i} correspond to the means of the control,
 stressor A, stressor B, and stressors A and B treatments respectively.

 ---

 A \emph{synergistic} classification is assigned where:

 an effect size is positive \emph{and} \emph{Expected} is greater than zero.

 or

 an effect size is negative \emph{and} \emph{Expected} is less than zero.

 ---

 An \emph{antagonistic} classification is assigned where:

 an effect size is negative \emph{and both} \emph{Expected and Observed} are greater than zero.

 or

 an effect size is positive \emph{and both} \emph{Expected and Observed} are less than zero.

 ---

 A \emph{reversal} classification is assigned where:

 an effect size is negative \emph{and} \emph{Expected} is greater than zero but \emph{Observed} is less than zero.

 or

 an effect size is positive \emph{and} \emph{Expected} is less than zero but \emph{Observed} is greater than zero.


 If \emph{assign_reversals} = FALSE, then where a reversal would be assigned using the above method, it is simply assigned
 an \emph{antagonistic} classification instead.

---

---

 If \emph{remove_directionality} = TRUE, then the following method is implemented.

 Where an interaction has an \emph{Expected} value <0, the corresponding effect size
 for this interaction has its polarity changed (i.e., ES = -ES).

 Confidence intervals for these altered effect sizes are likewise updated.

 This method is most likely required for meta-analyses having previously been implemented in
 several previous analyses (see \emph{Burgess et al. (2021)}; \emph{Jackson et al. (2016)};
 \emph{Seifert et al. (2020)}; \emph{Siviter et al. (2021)})
}
\description{
Using the output from either the \code{\link{effect_size_additive}} or \code{\link{effect_size_multiplicative}} functions,
interactions can be assigned classifications (e.g., antagonisms or synergisms) based on the frameworks used by
\emph{Burgess et al. (2021)}, \emph{Jackson et al. (2016)}, or \emph{Orr et al. (2020)}.
}
\examples{
classify_interactions(effect_size_dataframe = df,
                   assign_reversals = TRUE,
                   remove_directionality = TRUE)

}
\references{
Burgess, B. J., Purves, D., Mace, G., & Murrell, D. J. (2021). Classifying ecosystem stressor interactions: Theory highlights the data limitations of the additive null model and the difficulty in revealing ecological surprises. \emph{Global Change Biology}.

Jackson, M. C., Loewen, C. J., Vinebrooke, R. D., & Chimimba, C. T. (2016). Net effects of multiple stressors in freshwater ecosystems: a meta-analysis. \emph{Global Change Biology}, 22(1), 180-189.

Orr, J. A., Vinebrooke, R. D., Jackson, M. C., Kroeker, K. J., Kordas, R. L., Mantyka-Pringle, C., ... & Piggott, J. J. (2020). Towards a unified study of multiple stressors: divisions and common goals across research disciplines. \emph{Proceedings of the Royal Society B}, 287(1926), 20200421.

Seifert, M., Rost, B., Trimborn, S., & Hauck, J. (2020). Meta-analysis of multiple driver effects on marine phytoplannkton highlights modulating role of pCO2. \emph{Global Change Biology}, 26(12), 6787-6804.

Siviter, H., Bailes, E. J., Martin, C. D., Oliver, T. R., Koricheva, J., Leadbeater, E., & Brown, M. J. (2021). Agrochemicals interact synergistically to increase bee mortality. \emph{Nature}, 596(7872), 389-392.
}
